Reffactor to use adapter pattern
