+++ beacon_node/beacon_chain/src/canonical_head.rs
@@ -373,7 +373,7 @@ impl<T: BeaconChainTypes> CanonicalHead<T> {
     // TODO(gloas) just a stub for now, implement this once we have fork choice.
     /// Returns true if the payload for this block is canonical according to fork choice
     /// Returns an error if the block root doesn't exist in fork choice.
-    fn block_has_canonical_payload(&self, _root: &Hash256) -> Result<bool, Error> {
+    pub fn block_has_canonical_payload(&self, _root: &Hash256) -> Result<bool, Error> {
         Ok(true)
     }
 
@@ -427,20 +427,6 @@ impl<T: BeaconChainTypes> CanonicalHead<T> {
     }
 }
 
-// TODO(gloas) extend this trait to support methods needed by envelope_verification
-// so that its unit testable.
-pub trait CanonicalHeadReader: Send + Sync {
-    fn block_has_canonical_payload(&self, root: &Hash256) -> Result<bool, Error>;
-}
-
-impl<T: BeaconChainTypes> CanonicalHeadReader for BeaconChain<T> {
-    fn block_has_canonical_payload(&self, root: &Hash256) -> Result<bool, Error> {
-        self.canonical_head.block_has_canonical_payload(root)
-    }
-}
-
-// TODO(refactor) we can move any canonical head related read functions into the newly introduced trait
-// `CanonicalHeadReader`. This trait is useful for writing unit tests.
 impl<T: BeaconChainTypes> BeaconChain<T> {
     /// Contains the "best block"; the head of the canonical `BeaconChain`.
     ///
