+++ beacon_node/beacon_chain/src/lib.rs
@@ -24,7 +24,6 @@ pub mod envelope_times_cache;
 mod errors;
 pub mod events;
 pub mod execution_payload;
-pub mod execution_payload_envelope_streamer;
 pub mod fetch_blobs;
 pub mod fork_choice_signal;
 pub mod graffiti_calculator;
@@ -44,6 +43,7 @@ pub mod observed_block_producers;
 pub mod observed_data_sidecars;
 pub mod observed_operations;
 mod observed_slashable;
+pub mod payload_envelope_streamer;
 pub mod payload_envelope_verification;
 pub mod pending_payload_envelopes;
 pub mod persisted_beacon_chain;
